fix: shcedule days_of_week type validation 에러 해결

+++ sql/init.sql
@@ -144,7 +144,7 @@ CREATE TABLE schedule (
     start_date DATE NOT NULL,
     end_date DATE NOT NULL,
     time TIME DEFAULT NULL,
-    days_of_week INT DEFAULT NULL,
+    days_of_week TINYINT  DEFAULT NULL,
     alarm ENUM('TEN_MINUTE', 'ONE_HOUR', 'ONE_DAY') DEFAULT NULL,
     location VARCHAR(255) DEFAULT NULL,
     memo VARCHAR(255) DEFAULT NULL,
